Reaaally rough vision tracking rotation
Attempts of using the NavX angle as the PID Input with the angle being
read from the limelight being input into the
SimpleStupidPIDDriveRotateWithVisionCommand ended up with the robot
rotating to a setpoint that is inconsistent with the angle readout of
the limelight. Made the PID Input the angle of the limelight and was
able to get the PID rotate to land within 1.5 degrees of the target, but
it needs heavy tuning and potentially adding in Feedforward gain because
it is only reaching the point after a couple of seconds of the I-gain in
effect.